Implement routing in applicating
