+++ src/app/add-edit/add-edit.component.html
@@ -1,4 +1,4 @@
-<div id="overlay" *ngIf="addEditEnabled">
+<div id="overlay">
     <div class="box overlay-background">
         <div class="modal-card">
             <header class="modal-card-head border-background">
