+++ src/app/app-routing.module.ts
@@ -1,7 +1,13 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { AddEditComponent } from './add-edit/add-edit.component';
+import { PokemonSearchComponent } from './pokemon-search/pokemon-search.component';
 
-const routes: Routes = [];
+const routes: Routes = [
+  { path: 'add-edit', component: AddEditComponent },
+  { path: 'pokemon-search', component: PokemonSearchComponent },
+  { path: '**', redirectTo: 'pokemon-search' }
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
