+++ src/app/pokemon-search/pokemon-search.component.html
@@ -1,0 +1,9 @@
+ <!-- Filter Menu -->
+ <app-filter-menu (onFilterUpdate)="getFilterUpdate($event)" (onDisplayShortList)="displayShortList()"></app-filter-menu>
+
+ <!-- Table to Display Pokemon Matching Query -->
+ <div class="pt-4 pb-5">
+     <div class="table-container is-flex is-justify-content-center query-table-container">
+         <app-query-table [pkmnList]="pkmnList" [queryProcessing]="queryProcessing" (onAddToShortlist)="addToShortlist($event)" (onRemoveFromShortList)="removeFromShortList($event)"></app-query-table>
+     </div>
+ </div>
